Work around MinGW headers not having <winapifamily.h>
MinGW-w64 headers have a version of the winapifamily.h header file, but
MinGW headers (from the MinGW.OSDN project, formerly MinGW.org) do not.
To work around that, winapifamily.h is added for a Windows targeting
toolchain, only if it's not MinGW (based on
https://sourceforge.net/p/predef/wiki/Compilers/).

Simply not including winapifamily.h results in "error: missing binary
operator before token '('" for WINAPI_FAMILY_PARTITION() invocations.
WINAPI_FAMILY_PARTITION and WINAPI_PARTITION_DESKTOP are #defined so that
it compiles correctly.